requirements.txt
asn1crypto==0.24.0
astroid==2.0
autopep8==1.4.2
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
future==0.17.0
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pycodestyle==2.4.0
pycparser==2.19
pylint==2.0.0
python-telegram-bot==11.1.0
six==1.11.0
urllib3==1.24
virtualenv==16.0.0
wrapt==1.10.11
